Added the Create Image button
git-svn-id: https://svn.java.net/svn/swingx~svn/trunk@1785 1312f61e-266d-0410-97fa-c3b71232c9ac

+++ src/java/org/jdesktop/swingx/image/GaussianBlurFilter.java
@@ -67,7 +67,7 @@ public class GaussianBlurFilter extends AbstractFilter {
      *
      * @return the radius of the blur
      */
-    public float getRadius() {
+    public int getRadius() {
         return radius;
     }
